requirements.txt
amqp==5.1.1  
APScheduler==3.9.1.post1
asgiref==3.5.2
asn1crypto==1.5.1
async-timeout==4.0.2    
billiard==3.6.4.0  
boto3==1.26.51
botocore==1.29.51
backports.zoneinfo;python_version<"3.9"
celery==5.1.2
certifi==2022.9.24
cffi==1.15.1 
charset-normalizer==2.1.1
click==7.1.2    
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3   
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
django-filter==22.1 
drf-yasg==1.21.4
et-xmlfile==1.1.0
filelock==3.8.0
google-api-core==2.11.0
google-api-python-client==2.75.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
greenlet==2.0.2
httplib2==0.21.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mysql-connector-python==8.0.32
numpy==1.23.5
oauth2client==4.1.3
openpyxl==3.0.10
oscrypto==1.3.0
packaging==21.3
pandas==1.5.1
pip==22.3
prompt-toolkit==3.0.32
protobuf==3.20.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
pycryptodomex==3.15.0
PyJWT==1.7.1
pyodbc==4.0.35
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
requests==2.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
setuptools==65.3.0
six==1.16.0
snowflake-connector-python==2.8.2
SQLAlchemy==2.0.1
sqlparse==0.4.3
teradatasql==17.20.0.13
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wheel==0.37.1
wrapt==1.14.0

